requirements.txt
gql
python-dotenv
pandas